optimize + nits + fix corect train_on_completions
