+++ mlx_vlm/trainer/__init__.py
@@ -1,9 +1,11 @@
 from .lora import LoRaLayer, replace_lora_with_linear
-from .trainer import Dataset, TrainingArgs, Colors, train, save_adapter
+from .trainer import TrainingArgs, train, save_adapter
+from .datasets import Dataset
 from .utils import (
     apply_lora_layers,
     count_parameters,
     find_all_linear_names,
     get_peft_model,
     print_trainable_parameters,
+    Colors,
 )
